+++ src/commands/hardis/project/generate/flow-git-diff.ts
@@ -1,10 +1,9 @@
 /* jscpd:ignore-start */
 import { SfCommand, Flags } from '@salesforce/sf-plugins-core';
-import { Messages, SfError } from '@salesforce/core';
+import { Messages } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
 import c from "chalk";
-import * as Diff from "diff";
-import fs from "fs-extra";
+
 import * as path from "path";
 import {
   ensureGitRepository,
@@ -14,11 +13,9 @@ import {
 } from '../../../../common/utils/index.js';
 import { prompts } from '../../../../common/utils/prompts.js';
 import moment from 'moment';
-import { getReportDirectory } from '../../../../config/index.js';
-import { generateMarkdownFileWithMermaid, getMermaidExtraClasses } from '../../../../common/utils/mermaidUtils.js';
+import { generateFlowVisualGitDiff } from '../../../../common/utils/mermaidUtils.js';
 import { MetadataUtils } from '../../../../common/metadata-utils/index.js';
 import { WebSocketClient } from '../../../../common/websocketClient.js';
-import { parseFlow } from '../../../../common/utils/flowVisualiser/flowParser.js';
 
 Messages.importMessagesDirectoryFromMetaUrl(import.meta.url);
 const messages = Messages.loadMessages('sfdx-hardis', 'org');
@@ -112,47 +109,7 @@ Run \`npm install @mermaid-js/mermaid-cli --global\`
     const commitAfter = commitSelectRes.after;
     const commitBefore = commitSelectRes.before;
 
-    const mermaidMdBefore = await this.buildMermaidMarkdown(commitBefore);
-    const mermaidMdAfter = await this.buildMermaidMarkdown(commitAfter);
-
-    if (this.debugMode) {
-      uxLog(this, c.grey("FLOW DOC BEFORE:\n" + mermaidMdBefore) + "\n");
-      uxLog(this, c.grey("FLOW DOC AFTER:\n" + mermaidMdAfter) + "\n");
-    }
-
-    const flowDiffs = Diff.diffLines(mermaidMdBefore, mermaidMdAfter);
-    uxLog(this, JSON.stringify(flowDiffs, null, 2));
-
-    const mixedLines: any[] = [];
-    for (const line of flowDiffs) {
-      if (line.added) {
-        mixedLines.push(...line.value.split(/\r?\n/).map(lineSplit => { return ["added", lineSplit] }));
-      }
-      else if (line.removed) {
-        mixedLines.push(...line.value.split(/\r?\n/).map(lineSplit => { return ["removed", lineSplit] }));
-      }
-      else {
-        mixedLines.push(...line.value.split(/\r?\n/).map(lineSplit => { return ["unchanged", lineSplit] }));
-      }
-    }
-    // uxLog(this, JSON.stringify(mixedLines, null, 2));
-    const compareMdLines: string[] = [];
-    this.buildFinalCompareMarkdown(mixedLines, compareMdLines, false);
-
-    // Write markdown with diff in a file
-    const reportDir = await getReportDirectory();
-    await fs.ensureDir(path.join(reportDir, "flow-diff"));
-    const diffMdFile = path.join(reportDir, 'flow-diff', `${this.flowLabel}_${moment().format("YYYYMMDD-hhmmss")}.md`);
-    await fs.writeFile(diffMdFile, compareMdLines.join("\n"));
-    if (this.debugMode) {
-      await fs.copyFile(diffMdFile, diffMdFile + ".mermaid.md");
-    }
-
-    // Generate final markdown with mermaid SVG
-    const finalRes = await generateMarkdownFileWithMermaid(diffMdFile);
-    if (finalRes) {
-      uxLog(this, c.green(`Successfull generated visual git diff for flow: ${diffMdFile}`));
-    }
+    const diffMdFile = await generateFlowVisualGitDiff(this.flowFile, commitBefore, commitAfter, this.debugMode)
 
     // Open file in a new VsCode tab if available
     WebSocketClient.requestOpenFile(path.relative(process.cwd(), diffMdFile));
@@ -161,79 +118,6 @@ Run \`npm install @mermaid-js/mermaid-cli --global\`
     return {
       diffMdFile: diffMdFile
     };
-  }
-
-  private buildFinalCompareMarkdown(mixedLines: any[], compareMdLines, isMermaid) {
-    if (mixedLines.length === 0) {
-      return;
-    }
-    // Take line to process
-    const [status, currentLine] = mixedLines.shift();
-    // Update mermaid state
-    if (isMermaid === false && currentLine.includes("```mermaid")) {
-      isMermaid = true;
-    } else if (isMermaid === true && currentLine.includes("```")) {
-      compareMdLines.push(...getMermaidExtraClasses().split("\n"));
-      isMermaid = false;
-    }
-    let styledLine = currentLine;
-    // Remove next diff line if not relevant
-    if (styledLine.startsWith("|") && mixedLines.length > 1 && mixedLines[0][1] === '' && mixedLines[1][1].startsWith("|")) {
-      mixedLines.shift();
-    }
-
-    // Tables lines
-    if (!isMermaid && status === "removed" && styledLine.startsWith("|")) {
-      styledLine = "|" + styledLine.split("|").filter(e => e !== "").map((col: string) => `<span style="background-color: red;">${col}</span>`).join("|") + "|";
-    }
-    else if (!isMermaid && status === "added" && styledLine.startsWith("|")) {
-      styledLine = "|" + styledLine.split("|").filter(e => e !== "").map((col: string) => `<span style="background-color: green;">${col}</span>`).join("|") + "|";
-    }
-    // Normal lines
-    else if (!isMermaid && status === "removed" && styledLine !== "") {
-      styledLine = `<span style="background-color: red;">${styledLine}</span>`;
-    }
-    else if (!isMermaid && status === "added" && styledLine !== "") {
-      styledLine = `<span style="background-color: green;">${styledLine}</span>`;
-    }
-    // Boxes lines
-    else if (isMermaid === true && status === "removed" && currentLine.split(":::").length === 2) {
-      styledLine = styledLine + "Removed"
-    }
-    else if (isMermaid === true && status === "added" && currentLine.split(":::").length === 2) {
-      styledLine = styledLine + "Added"
-    }
-    else if (isMermaid === true && currentLine.includes(":::")) {
-      // Detect if link line does not change, but its content did
-      const splits = currentLine.split(/[[({]/);
-      if (splits.length > 1) {
-        const boxName = splits[0];
-        const changed = mixedLines.filter(([lineStatus, line]) => { return line.startsWith(`click ${boxName}`) && ["added", "changed"].includes(lineStatus) }).length;
-        if (changed > 0) {
-          styledLine = styledLine + "Changed"
-        }
-      }
-    }
-    // Link lines
-    else if (isMermaid === true && status === "removed" && currentLine.includes('-->')) {
-      styledLine = styledLine.replace("-->", "-.->") + ":::removedLink"
-    }
-    else if (isMermaid === true && status === "added" && currentLine.includes('-->')) {
-      styledLine = styledLine.replace("-->", "==>") + ":::addedLink"
-    }
-    compareMdLines.push(styledLine);
-    // Continue processing next lines
-    this.buildFinalCompareMarkdown(mixedLines, compareMdLines, isMermaid)
-  }
-
-  private async buildMermaidMarkdown(commit) {
-    const flowXml = await git().show([`${commit}:${this.flowFile}`]);
-    try {
-      const flowDocGenResult = await parseFlow(flowXml, 'mermaid', { outputAsMarkdown: true });
-      return flowDocGenResult.uml;
-    } catch (err: any) {
-      throw new SfError(`Unable to build Graph for flow ${this.flowFile}: ${err.message}`)
-    }
   }
 
 }
